requirements.txt
pandas==2.3.3
requests==2.32.5
openpyxl==3.1.5
python-dotenv==1.2.1
discord.py>=2.3
matplotlib>=3.8
pytz>=2024.1
